Converted SplineWriter to use floats instead of doubles for co-efficients. Stopped T2KRW talking during reconfigure stage. Added estimated time remaining statement when generating splines.

+++ src/Splines/SplineWriter.h
@@ -19,7 +19,8 @@ class SplineWriter : public SplineReader {
 
 
   int fNCoEff;
-  double* fCoEffStorer;
+  //  double* fCoEffStorer;
+  float* fCoEffStorer;
 
   std::vector< std::vector<double> > fParVect;
   std::vector< int > fSetIndex;
